[improved] serialization and remove useless class
