+++ backend/src/main/kotlin/pl/akai/fillist/security/models/AccessTokenResponseBody.kt
@@ -1,12 +1,9 @@
-package pl.akai.fillist.security.models
-
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class AccessTokenResponseBody(
+class AccessTokenResponseBody(
     val accessToken: String,
     val tokenType: String,
     val expiresIn: Int,
     val refreshToken: String,
-) {
-}
+)
