+++ backend/src/main/kotlin/pl/akai/fillist/security/models/RefreshTokenRequestBody.kt
@@ -1,0 +1,9 @@
+package pl.akai.fillist.security.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RefreshTokenRequestBody(
+    private val grantType: String,
+    private val refreshToken: String
+)
