+++ backend/src/main/kotlin/pl/akai/fillist/security/service/Oauth2SsoService.kt
@@ -1,9 +1,9 @@
 package pl.akai.fillist.security.service
 
+import AccessTokenResponseBody
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import pl.akai.fillist.security.models.AccessTokenRequestBody
-import pl.akai.fillist.security.models.AccessTokenResponseBody
 import pl.akai.fillist.security.models.AuthorizationCodeUrlResponseBody
 import reactor.core.publisher.Mono
